Fix existing EIP matching
If the user specifies private_ip_address, allow for
the check of EIP existance to match that specific address.

Previously if you wanted to attach an EIP to a specific
private-ip of an ENI WHILE there is an existing EIP
attached to some other private-ip on the same ENI,
the module would not change anything.